pyproject.toml
[tool.poetry]
name = "rl-experiments"
version = "0.1.0"
description = "A protocol for RL experiments"
authors = ["Roberto Cipollone <cipollone.rt@gmail.com>"]
include = ["**/*.yaml"]

[tool.poetry.dependencies]
python = "^3.7"
PyYAML = "^5.4.1"

[tool.poetry.dev-dependencies]
ipython = "^7.22.0"
pytest = "^6.2.3"
mypy = "^0.812"
isort = "^5.8.0"
black = "^20.8b1"
jedi = "^0.18.0"
flake8 = "^3.9.0"
tox = "^3.23.0"

[tool.isort]
profile = "black"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
